Fix comments, restructure exports
