+++ test/Hotels/HotelsParser.test.js
@@ -1,7 +1,7 @@
 
 const assert = require('assert');
 const fs = require('fs');
-const { Parser } = require('../../src/Request/uapi-parser');
+const Parser = require('../../src/Request/uapi-parser');
 const hotelsParser = require('../../src/Services/Hotels/HotelsParser');
 
 const xmlFolder = `${__dirname}/../FakeResponses/Hotels`;
